+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   the wheel distribution, but users invoking pip with `--no-binary :all:` will continue having
   installs based on the source distribution.
 
+- Return mapping information for terraform conversions (https://github.com/pulumi/pulumi-kubernetes/pull/2457)
+
 ## 4.1.1 (August 23, 2023)
 
 - Revert the switch to pyproject.toml and wheel-based PyPI publishing as it impacts users that run pip with --no-binary